src/docker-rapid-development-tricks/requirements.txt
redis==5.0.7
psycopg2-binary==2.9.9
pandas==2.2.2
scikit-learn==1.5.1
flask==3.0.3
